sha1dc: add collision-detecting sha1 implementation
This is the current state of the sha1 collision detection from

    https://github.com/cr-marcstevens/sha1collisiondetection.git

pared down to just the minimal implementation (in lib/) and the
LICENSE.txt file.

Top-of-tree of the source tree is commit 007905a93c97 ("Merge pull
request #20 from cr-marcstevens/feature/performance").  Thanks to both
Marc and Dan for making the code fit our needs by doing both
optimization work, cutting down on the object size, and doing some
syntactic changes to work better with git.

The license of the sha1dc code is the MIT license, which is obviously
compatible with the GPLv2 of git.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>