http: add http.keepinvalidcredentials config
If a Git HTTP server responds with 401 or 407, then Git tells the
credential helper to reject and delete the credentials. In general
this is good.

However, in certain automation environments it is not desired to remove
credentials automatically. This is in particular the case if credentials
are only invalid temporarily (e.g. because of problems in the server's
authentication backend).

Therefore, add the config "http.keepinvalidcredentials" which tells
Git to keep invalid credentials if set to "true".

It was considered to disable the credential reset in credential.c
directly. This approach was not chosen as it could be confusing to a
caller of credential_reject() if the function does not do what its name
says.

The Git-Credential-Manager-for-Windows already implementens a similar
mechansim using the environment variable GCM_PRESERVE_CREDS [1]. This
solution aims to enable that feature for all credential helper
implementations.

[1] https://github.com/Microsoft/Git-Credential-Manager-for-Windows/issues/58#issuecomment-259544029

Signed-off-by: Lars Schneider <larsxschneider@gmail.com>